\name{dfClasses}
\alias{dfClasses}
\title{
dfClasses
}

\description{
display classes of the columns of a data.frame
}

\usage{
dfClasses(x)
}

\arguments{
  \item{x}{ a data.frame object}
}

\section{Side Effects}{
data.frame column numbers, names, and classes are printed
}

\author{
Guy Yollin, \url{http://www.r-programming.org}
}

\examples{
dfClasses(stackloss)
}

