\name{facmod-package}
\alias{facmod-package}
\alias{facmod}
\docType{package}

\title{
Factor models for asset returns
}

\description{
Fit cross-sectional time series factor models via classical or robust linear regression
}

\details{
\tabular{ll}{
Package: \tab facmod\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2011-05-17\cr
License: \tab GNU GPL v2\cr
LazyLoad: \tab yes\cr
Depends: \tab methods\cr
}
~~ An overview of how to use the package, including the most important functions ~~
}

\author{
Guy Yollin, \url{http://www.r-programming.org}\cr
(based on S-PLUS code developed by Christopher G. Green, Eric Aldrich, and Yindeng Jiang of the University of Washington) \cr
Maintainer: Guy Yollin <gyollin@r-programming.org>
}

\examples{
~~ simple examples of the most important functions ~~
}
